% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/divide.R
\name{divide}
\alias{divide}
\title{Divide two numbers with a check for division by zero}
\usage{
divide(x, y)
}
\arguments{
\item{x}{The numerator}

\item{y}{The denominator}
}
\value{
The result of x divided by y
}
\description{
This function takes two numbers and divides the first number by the second number.
If the second number is 0, it outputs an error message.
}
\examples{
divide(6, 2)
# Returns: 3

divide(4, 0)
# Outputs an error message: "You cannot divide by 0."

}
